+++ __init__.py
@@ -1,4 +1,4 @@
 from .converter import MicronConverter, default_url_resolver, UrlResolver
 
 __all__ = ["MicronConverter", "default_url_resolver", "UrlResolver"]
-__version__ = "1.0.0"
+__version__ = "1.0.2"